feat: canvas group으로 이동, 신규 setting 모듈 생성
