+++ app/build.gradle.kts
@@ -47,17 +47,22 @@ dependencies {
     implementation(projects.feature.camera.impl)
     implementation(projects.feature.gallery.api)
     implementation(projects.feature.gallery.impl)
-    implementation(projects.feature.canvas.api)
-    implementation(projects.feature.canvas.impl)
     implementation(projects.feature.splash.api)
     implementation(projects.feature.splash.impl)
 
+    implementation(projects.feature.app.setting.api)
+    implementation(projects.feature.app.setting.impl)
+
+    implementation(projects.feature.groups.canvas.api)
+    implementation(projects.feature.groups.canvas.impl)
     implementation(projects.feature.groups.enter.api)
     implementation(projects.feature.groups.enter.impl)
     implementation(projects.feature.groups.home.api)
     implementation(projects.feature.groups.home.impl)
     implementation(projects.feature.groups.list.api)
     implementation(projects.feature.groups.list.impl)
+    implementation(projects.feature.groups.setting.api)
+    implementation(projects.feature.groups.setting.impl)
 
     implementation(projects.domain)
     implementation(projects.data)
