+++ feature/app/setting/api/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    alias(libs.plugins.parfait.module.feature.api)
+}
+
+android {
+    namespace = "com.teamyg.parfait.feature.app.setting.api"
+}
