+++ feature/app/setting/api/src/main/kotlin/com/teamyg/parfait/feature/app/setting/api/NavKeyAppSetting.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.app.setting.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NavKeyAppSetting : NavKey
