+++ feature/app/setting/impl/src/main/kotlin/com/teamyg/parfait/feature/app/setting/impl/route/AppSettingRoute.kt
@@ -1,0 +1,13 @@
+package com.teamyg.parfait.feature.app.setting.impl.route
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.teamyg.parfait.core.navigation.Navigator
+
+@Composable
+internal fun AppSettingRoute(
+    navigator: Navigator,
+    modifier: Modifier = Modifier,
+) {
+    // TODO impl
+}
