--- feature/canvas/api/src/main/java/com/teamyg/parfait/feature/canvas/api/NavKeyCanvasEdit.kt
@@ -1,7 +1,0 @@
-package com.teamyg.parfait.feature.canvas.api
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class NavKeyCanvasEdit(val imageUri: String) : NavKey
