--- feature/canvas/api/src/main/java/com/teamyg/parfait/feature/canvas/api/NavKeyCanvasImageAdd.kt
@@ -1,7 +1,0 @@
-package com.teamyg.parfait.feature.canvas.api
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object NavKeyCanvasImageAdd : NavKey
