--- feature/canvas/impl/src/main/java/com/teamyg/parfait/feature/canvas/impl/navigation/NavigationModule.kt
@@ -1,20 +1,0 @@
-package com.teamyg.parfait.feature.canvas.impl.navigation
-
-import androidx.navigation3.runtime.EntryProviderScope
-import androidx.navigation3.runtime.NavKey
-import com.teamyg.parfait.core.navigation.Navigator
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityRetainedComponent
-import dagger.multibindings.IntoSet
-
-@Module
-@InstallIn(ActivityRetainedComponent::class)
-object NavigationModule {
-    @IntoSet
-    @Provides
-    fun provideFeatureCanvasEntryBuilder(): EntryProviderScope<NavKey>.(Navigator) -> Unit = {
-        featureCanvasEntryBuilder(navigator = it)
-    }
-}
