+++ feature/groups/canvas/api/src/main/AndroidManifest.xml
