+++ feature/groups/canvas/api/src/main/kotlin/com/teamyg/parfait/feature/groups/canvas/api/NavKeyCanvasImageAdd.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.groups.canvas.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NavKeyCanvasImageAdd : NavKey
