+++ feature/groups/canvas/impl/src/main/AndroidManifest.xml
