+++ feature/groups/canvas/impl/src/main/kotlin/com/teamyg/parfait/feature/groups/canvas/impl/navigation/EntryBuilder.kt
@@ -1,0 +1,50 @@
+package com.teamyg.parfait.feature.groups.canvas.impl.navigation
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.ui.Modifier
+import androidx.navigation3.runtime.EntryProviderScope
+import androidx.navigation3.runtime.NavKey
+import com.teamyg.parfait.feature.groups.canvas.api.NavKeyCanvasEdit
+import com.teamyg.parfait.feature.groups.canvas.api.NavKeyCanvasImageAdd
+import com.teamyg.parfait.feature.groups.canvas.api.NavKeyCanvasImageSelect
+import com.teamyg.parfait.feature.groups.canvas.impl.route.CanvasEditRoute
+import com.teamyg.parfait.feature.groups.canvas.impl.route.CanvasImageAddRoute
+import com.teamyg.parfait.feature.groups.canvas.impl.route.CanvasImageSelectRoute
+import com.teamyg.parfait.core.navigation.Navigator
+
+fun EntryProviderScope<NavKey>.featureCanvasEntryBuilder(navigator: Navigator) {
+    entry<NavKeyCanvasImageAdd> {
+        Scaffold { innerPadding ->
+            CanvasImageAddRoute(
+                navigator = navigator,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(innerPadding),
+            )
+        }
+    }
+    entry<NavKeyCanvasEdit> { navKey ->
+        Scaffold { innerPadding ->
+            CanvasEditRoute(
+                imageUri = navKey.imageUri,
+                navigator = navigator,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(innerPadding),
+            )
+        }
+    }
+
+    entry<NavKeyCanvasImageSelect> {
+        Scaffold { innerPadding ->
+            CanvasImageSelectRoute(
+                navigator = navigator,
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(innerPadding),
+            )
+        }
+    }
+}
