+++ feature/groups/canvas/impl/src/main/kotlin/com/teamyg/parfait/feature/groups/canvas/impl/navigation/NavigationModule.kt
@@ -1,0 +1,20 @@
+package com.teamyg.parfait.feature.groups.canvas.impl.navigation
+
+import androidx.navigation3.runtime.EntryProviderScope
+import androidx.navigation3.runtime.NavKey
+import com.teamyg.parfait.core.navigation.Navigator
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ActivityRetainedComponent
+import dagger.multibindings.IntoSet
+
+@Module
+@InstallIn(ActivityRetainedComponent::class)
+object NavigationModule {
+    @IntoSet
+    @Provides
+    fun provideFeatureCanvasEntryBuilder(): EntryProviderScope<NavKey>.(Navigator) -> Unit = {
+        featureCanvasEntryBuilder(navigator = it)
+    }
+}
