+++ feature/groups/canvas/impl/src/main/res/drawable/nukkiii.png
