+++ feature/groups/setting/api/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    alias(libs.plugins.parfait.module.feature.api)
+}
+
+android {
+    namespace = "com.teamyg.parfait.feature.groups.setting.api"
+}
