+++ feature/groups/setting/api/src/main/kotlin/com/teamyg/parfait/feature/groups/setting/api/NavKeyGroupSetting.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.groups.setting.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NavKeyGroupSetting : NavKey
